% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrmatrix.R
\name{corrmatrix}
\alias{corrmatrix}
\title{A NBA Function}
\usage{
corrmatrix(year)
}
\arguments{
\item{year}{The specific year/season}
}
\description{
This function returns the correlation matrix for all numeric variables in the NBA data for a given year.
}
\examples{
corrmatrix()
}
\keyword{correlation}
\keyword{matrix}
\keyword{the}
