% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nba_data.R
\name{do_get_nba_data}
\alias{do_get_nba_data}
\title{Actually get the data}
\usage{
do_get_nba_data()
}
\value{
A Data Frame containing the NBA Data
}
\description{
This function gets the data for nba players
}
\examples{
do_get_nba_data()
}

