% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{espn_nba_betting}
\alias{espn_nba_betting}
\title{Get ESPN NBA betting information}
\usage{
espn_nba_betting(game_id, save_raw = FALSE, raw_dir = "data/raw")
}
\arguments{
\item{game_id}{Game ID}

\item{save_raw}{Save raw JSON to disk}

\item{raw_dir}{Directory to store raw JSON}
}
\value{
A named list of data frames: pickcenter, againstTheSpread, predictor
}
\description{
Get ESPN NBA betting information
}
