% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{espn_nba_find_boxscore_game_id}
\alias{espn_nba_find_boxscore_game_id}
\title{Find a recent completed game with boxscore data}
\usage{
espn_nba_find_boxscore_game_id(days_back = 30, end_date = Sys.Date())
}
\arguments{
\item{days_back}{Number of days to look back}

\item{end_date}{Date to end the search (defaults to Sys.Date())}
}
\value{
A game_id or NULL if none found
}
\description{
Find a recent completed game with boxscore data
}
