% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{espn_nba_game_has_boxscore}
\alias{espn_nba_game_has_boxscore}
\title{Check if a game has boxscore data available}
\usage{
espn_nba_game_has_boxscore(game_id)
}
\arguments{
\item{game_id}{Game ID}
}
\value{
TRUE if boxscore data is available, FALSE otherwise
}
\description{
Check if a game has boxscore data available
}
