% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{espn_nba_player_box}
\alias{espn_nba_player_box}
\title{Get ESPN NBA player box scores}
\usage{
espn_nba_player_box(game_id, save_raw = FALSE, raw_dir = "data/raw")
}
\arguments{
\item{game_id}{Game ID}

\item{save_raw}{Save raw JSON to disk}

\item{raw_dir}{Directory to store raw JSON}
}
\value{
A player boxscore data frame
}
\description{
Get ESPN NBA player box scores
}
