% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{espn_nba_scoreboard}
\alias{espn_nba_scoreboard}
\title{Get ESPN NBA scoreboard}
\usage{
espn_nba_scoreboard(dates, save_raw = FALSE, raw_dir = "data/raw")
}
\arguments{
\item{dates}{Date(s) in YYYYMMDD format}

\item{save_raw}{Save raw JSON to disk}

\item{raw_dir}{Directory to store raw JSON}
}
\value{
A data frame of games
}
\description{
Get ESPN NBA scoreboard
}
