% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{espn_nba_scrape_game}
\alias{espn_nba_scrape_game}
\title{Scrape summary-based data for a single game}
\usage{
espn_nba_scrape_game(game_id, save_raw = TRUE, raw_dir = "data/raw")
}
\arguments{
\item{game_id}{Game ID}

\item{save_raw}{Save raw JSON to disk}

\item{raw_dir}{Directory to store raw JSON}
}
\value{
A named list with team, player, betting data frames
}
\description{
Scrape summary-based data for a single game
}
