% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{espn_nba_scraped_game_ids}
\alias{espn_nba_scraped_game_ids}
\title{Get game ids from scraped summary JSON files}
\usage{
espn_nba_scraped_game_ids(raw_dir = "data/raw")
}
\arguments{
\item{raw_dir}{Directory to store raw JSON}
}
\value{
Integer vector of scraped game ids
}
\description{
Get game ids from scraped summary JSON files
}
