% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{espn_nba_season_game_ids}
\alias{espn_nba_season_game_ids}
\title{Get ESPN NBA season game ids}
\usage{
espn_nba_season_game_ids(
  season,
  season_type = c("regular", "postseason", "all"),
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{season}{Season year}

\item{season_type}{One of "regular", "postseason", or "all"}

\item{start_date}{Optional season start date override}

\item{end_date}{Optional season end date override}
}
\value{
A data frame of games with ids
}
\description{
Get ESPN NBA season game ids
}
