% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{espn_nba_standings}
\alias{espn_nba_standings}
\title{Get ESPN NBA standings}
\usage{
espn_nba_standings(season = NULL, save_raw = FALSE, raw_dir = "data/raw")
}
\arguments{
\item{season}{Optional season year}

\item{save_raw}{Save raw JSON to disk}

\item{raw_dir}{Directory to store raw JSON}
}
\value{
A standings data frame
}
\description{
Get ESPN NBA standings
}
