% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{espn_nba_summary_raw}
\alias{espn_nba_summary_raw}
\title{Get ESPN NBA game summary raw JSON}
\usage{
espn_nba_summary_raw(game_id, save_raw = FALSE, raw_dir = "data/raw")
}
\arguments{
\item{game_id}{Game ID}

\item{save_raw}{Save raw JSON to disk}

\item{raw_dir}{Directory to store raw JSON}
}
\value{
A list containing the raw summary JSON
}
\description{
Get ESPN NBA game summary raw JSON
}
