% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raw.R
\name{espn_nba_team_box_from_raw}
\alias{espn_nba_team_box_from_raw}
\title{Parse ESPN NBA team box score from raw JSON}
\usage{
espn_nba_team_box_from_raw(raw_summary)
}
\arguments{
\item{raw_summary}{Summary JSON list}
}
\value{
A data frame of team box scores
}
\description{
Parse ESPN NBA team box score from raw JSON
}
