% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{espn_nba_valid_game_ids}
\alias{espn_nba_valid_game_ids}
\title{Get valid game ids for a date range with required data}
\usage{
espn_nba_valid_game_ids(
  dates,
  require = c("team", "player"),
  save_raw = FALSE,
  raw_dir = "data/raw",
  quiet = FALSE
)
}
\arguments{
\item{dates}{Vector of dates or a range in YYYYMMDD or YYYYMMDD-YYYYMMDD}

\item{require}{One or more of "summary", "team", "player", "betting"}

\item{save_raw}{Save raw JSON to disk}

\item{raw_dir}{Directory to store raw JSON}

\item{quiet}{Suppress progress messages}
}
\value{
A vector of game ids
}
\description{
Get valid game ids for a date range with required data
}
