% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_x_points.R
\name{get_last_x_points}
\alias{get_last_x_points}
\title{Get a Player's previous points}
\usage{
get_last_x_points(num_games)
}
\arguments{
\item{num_games}{how many games to provide the previous points for}
}
\value{
A dataframe containing the points data
}
\description{
This function returns a dataframe containing the points a player scored in a match
alongside the points he scored in the previous x matches
}
\examples{
get_last_x_points(6)
}

