% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nba_data.R
\name{get_nba_data}
\alias{get_nba_data}
\title{Get NBA Game Data}
\usage{
get_nba_data()
}
\value{
A Data Frame containing the NBA Data
}
\description{
This function allows you to get player by player game data from NBA Games
}
\examples{
get_nba_data()
}

