% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_form.R
\name{get_player_form}
\alias{get_player_form}
\title{Get Player form data}
\usage{
get_player_form(since, min_games = 6, min_av_points = 20)
}
\arguments{
\item{since}{date from which to look at a players form}

\item{min_games}{how many games should a player have in the period to be included}

\item{min_av_points}{what av points should a player have in the period to be included}
}
\value{
A plyr object containing the form data
}
\description{
This function allows you to get the form of all players for the period requested
}
\examples{
get_player_form("2016-03-01", 6, 20)
}

