% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getavg_season_age.R
\name{getavg_season_age}
\alias{getavg_season_age}
\title{A NBA Function}
\usage{
getavg_season_age(year)
}
\arguments{
\item{year}{The specific year/season}
}
\description{
This function returns the average age of NBA players for a specific year/season
}
\examples{
getavg_season_age()
}
\keyword{Average}
\keyword{age}
