% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBA.R
\docType{data}
\name{NBA}
\alias{NBA}
\title{Statistics of NBA players.}
\format{
A data frame with 24691 rows and 52 variables:
\describe{
\item{Year}{Year, int}
\item{Player}{Name of players, char}
...
}
}
\source{
\url{https://www.kaggle.com/drgilermo/nba-players-stats/}
}
\usage{
NBA
}
\description{
A dataset containing the position and other attributes of 24691 NBA players.
}
\keyword{datasets}
