% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrations.R
\name{nba_db_apply_migrations}
\alias{nba_db_apply_migrations}
\title{Apply schema migrations}
\usage{
nba_db_apply_migrations(con, from, to)
}
\arguments{
\item{con}{DBI connection}

\item{from}{Current schema version}

\item{to}{Target schema version}
}
\value{
TRUE invisibly
}
\description{
Apply schema migrations
}
