% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{nba_db_connect}
\alias{nba_db_connect}
\title{Connect to NBAData database}
\usage{
nba_db_connect(
  db = "data/sql/nbadata.sqlite",
  drv = c("sqlite", "postgres"),
  ...
)
}
\arguments{
\item{db}{Database path (sqlite) or connection string (postgres)}

\item{drv}{One of "sqlite" or "postgres"}

\item{...}{Additional DBI connection args}
}
\value{
A DBI connection
}
\description{
Connect to NBAData database
}
