% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{nba_db_disconnect}
\alias{nba_db_disconnect}
\title{Disconnect from NBAData database}
\usage{
nba_db_disconnect(con)
}
\arguments{
\item{con}{DBI connection}
}
\value{
TRUE invisibly
}
\description{
Disconnect from NBAData database
}
