% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrations.R
\name{nba_db_get_version}
\alias{nba_db_get_version}
\title{Get current schema version}
\usage{
nba_db_get_version(con)
}
\arguments{
\item{con}{DBI connection}
}
\value{
Integer schema version
}
\description{
Get current schema version
}
