% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{nba_db_init}
\alias{nba_db_init}
\title{Initialize NBAData database schema}
\usage{
nba_db_init(con, target_version = NBA_SCHEMA_VERSION)
}
\arguments{
\item{con}{DBI connection}

\item{target_version}{Schema version to migrate to}
}
\value{
TRUE invisibly
}
\description{
Initialize NBAData database schema
}
