% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastorr_helpers.R
\name{nbadata}
\alias{nbadata}
\alias{nbadata_del}
\alias{nbadata_version_current}
\alias{nbadata_versions}
\title{Download example data set}
\usage{
nbadata(version = NULL, path = NULL)

nbadata_versions(local = TRUE, path = NULL)

nbadata_version_current(local = TRUE, path = NULL)

nbadata_del(version, path = NULL)
}
\arguments{
\item{version}{Version number.  The default will load the most
recent version on your computer or the most recent version known
to the package if you have never downloaded the data before.
With \code{nbadata_del}, specifying \code{version=NULL} will
delete \emph{all} data sets.}

\item{path}{Path to store the data at.  If not given,
\code{datastorr} will use \code{rappdirs} to find the best place
to put persistent application data on your system.  You can
delete the persistent data at any time by running
\code{nbadata_del(NULL)} (or \code{nbadata_del(NULL, path)} if you
use a different path).}

\item{local}{Logical indicating if local or github versions should
be polled.  With any luck, \code{local=FALSE} is a superset of
\code{local=TRUE}.  For \code{nbadata_version_current}, if
\code{TRUE}, but there are no local versions, then we do check
for the most recent github version.}
}
\description{
Download the example data set from madams1/nbadata
 (\url{https://github.com/madams1/nbadata/})
}

