% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{nbadata_load_all}
\alias{nbadata_load_all}
\title{Load all nbadata into workspace}
\usage{
nbadata_load_all(path = "./")
}
\arguments{
\item{path}{A character string for the path to load nbadata files from.
Defaults to current working directory.}
}
\value{
Called for its side effects. Returns invisible NULL
}
\description{
Load all nbadata into workspace
}

