% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba-fpts.r
\docType{data}
\name{nbafpts}
\alias{nbafpts}
\title{Fantasy Points from Boxscores}
\format{Data frame with columns
\describe{
\item{game_date}{Date game played}
\item{team}{Team three-letter abbreviation}
\item{player_name}{Player first and last name}
\item{pos}{Player position during game}
\item{mpg}{Minutes per game}
\item{fpts_dk}{DraftKings fantasy points}
\item{fpts_fd}{FanDuel fantasy points}
}}
\source{
\url{https://www.kaggle.com/pablote/nba-enhanced-stats}
}
\usage{
nbafpts
}
\description{
Fantasy points for every game in the 2017-2018 NBA season. Calculated from boxscore data.
}
\examples{
  nbafpts
}
\keyword{datasets}
