% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldestplayer.R
\name{oldestplayer}
\alias{oldestplayer}
\title{A NBA Function}
\usage{
oldestplayer(year)
}
\arguments{
\item{year}{The specific year/season}
}
\description{
This function returns the oldest player in the NBA data for a given year. If multiple players have the same age, we return the first player by ascending name.
}
\examples{
oldestplayer()
}
\keyword{oldest}
\keyword{player}
\keyword{the}
