% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helper_espn_nba_team_box}
\alias{helper_espn_nba_team_box}
\title{Parse ESPN NBA Team Box}
\usage{
helper_espn_nba_team_box(resp)
}
\arguments{
\item{resp}{Response object from the ESPN NBA game summary endpoint}
}
\value{
Returns a data frame
}
\description{
Parse ESPN NBA Team Box
}
